+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/CourseLocationImageRepository.java
@@ -3,6 +3,6 @@ package com.moham.coursemores.repository;
 import com.moham.coursemores.domain.CourseLocationImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseLocationImageRepository extends JpaRepository<CourseLocationImage, Integer> {
-    void deleteByCourseLocationId(int courseLocationId);
+public interface CourseLocationImageRepository extends JpaRepository<CourseLocationImage, Long> {
+    void deleteByCourseLocationId(Long courseLocationId);
 }
