+++ BackEnd/SOA/src/main/java/com/moham/coursemores/repository/ThemeOfCourseRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ThemeOfCourseRepository extends JpaRepository<ThemeOfCourse, Integer> {
-    List<ThemeOfCourse> findByCourseId(int courseId);
-    void deleteByCourseId(int courseId);
+public interface ThemeOfCourseRepository extends JpaRepository<ThemeOfCourse, Long> {
+    List<ThemeOfCourse> findByCourseId(Long courseId);
+    void deleteByCourseId(Long courseId);
 }
