+++ BackEnd/SOA/src/main/java/com/moham/coursemores/service/UserService.java
@@ -3,7 +3,5 @@ package com.moham.coursemores.service;
 import com.moham.coursemores.dto.profile.UserSimpleInfoResDto;
 
 public interface UserService {
-    UserSimpleInfoResDto getUserSimpleInfo(int userId);
-
-
+    UserSimpleInfoResDto getUserSimpleInfo(Long userId);
 }
